Moved request tests to separated directory.
