+++ config/test.py
@@ -1,5 +1,5 @@
 SQLALCHEMY_DATABASE_URI = (
-    'postgresql://postgres:@localhost/bookbrainz_test'
+    'postgresql://bookbrainz:bookbrainz@localhost/bookbrainz_test'
 )
 
 REDIS_URL = 'redis://:@localhost:6379'
