+++ tests/requests_tests/get_list_testing.py
@@ -21,8 +21,8 @@ import uuid
 
 from flask_testing import TestCase
 
-from check_helper_functions import *
-from constants import *
+from ..check_helper_functions import *
+from ..constants import *
 
 
 class GetListTests(TestCase):
