+++ tests/requests_tests/put_testing.py
@@ -20,8 +20,8 @@ import logging
 
 from flask_testing import TestCase
 
-from check_helper_functions import *
-from constants import *
+from ..check_helper_functions import *
+from ..constants import *
 
 
 class PutTests(TestCase):
